When "Get Job Status" endpoint returns 200, it populates transactionTime with the job's completion timestamp. This is incorrect. That value should be set to the job's creation time.
