+++ eventlogger/src/main/java/gov/cms/ab2d/eventlogger/EventLogger.java
@@ -1,0 +1,5 @@
+package gov.cms.ab2d.eventlogger;
+
+public interface EventLogger {
+    void log(LoggableEvent event);
+}
