+++ eventlogger/src/main/java/gov/cms/ab2d/eventlogger/events/FileEvent.java
@@ -1,33 +1,23 @@
-package gov.cms.ab2d.eventlogger;
+package gov.cms.ab2d.eventlogger.events;
 
+import gov.cms.ab2d.eventlogger.LoggableEvent;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.time.OffsetDateTime;
-
 @Data
 @AllArgsConstructor
-public class FileEventLogger implements LoggableEvent {
+public class FileEvent extends LoggableEvent {
     public enum FileStatus {
         OPEN,
         CLOSE,
         DELETE
     }
-    // id
-    private Long id;
     // The file name involved
     private String fileName;
-    // The job it is associated with
-    private String jobId;
     // The new of the file
     private FileStatus status;
     // The size of the file if we're closing it
     private long fileSize;
     // The content hash so we can compare different files
     private String fileHash;
-
-    @Override
-    public boolean log(OffsetDateTime eventTime) {
-        return false;
-    }
 }
